research-pilot-ai/requirements.txt
groq
tavily-python
python-dotenv
streamlit
sentence-transformers
faiss-cpu
numpy